% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clinical.ranksum.R
\name{clinical.ranksum}
\alias{clinical.ranksum}
\title{Compute associations between gene set/module and patient clinical variables}
\usage{
clinical.ranksum(mixt.dat, mixt.ranksum, tissue, cohort.name = "all",
  corType = "p", nRuns = 10000, randomSeed = 12345, mc.cores = 2,
  verbose = 2)
}
\arguments{
\item{mixt.dat}{Data object from matched tissues.}

\item{mixt.ranksum}{output of sig.ranksum()}

\item{tissue}{character string that provides the name of tissue we want to
look at}

\item{cohort.name}{character string that provides the name of the patient
cohort to analyze, default is set to 'all'}

\item{corType}{a character string indicating which correlation coefficient
is to be computed. Default 'p' for pearson}

\item{nRuns}{number of permutations}

\item{randomSeed}{seed number for random number generation.
Default set as '1234'}

\item{mc.cores}{number of cores}

\item{verbose}{numerical. default > 0 show informational text on progress}
}
\value{
matrix of p-values determining significance of association between
clinical variables (in row) and gene sets/modules (in column)
}
\description{
Using ranksums to capture gene set/module expression, we can ask how gene
sets/modules in each tissue are differentially expressed
according to patient’s clinicopathological variables.
The type of the clinicopathological attribute (categorical or continuous)
determines the underlying statistical test.
Pearson correlation (Student asymptotic p-value) is used to test association
between a given module and continuous patient attributes (eg. age).
Analysis of Variance (ANOVA) is used to test association
between a given module and categorical patient attributes (eg. ER status).
}
