% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohort_boxplot.R
\name{cohort_boxplot}
\alias{cohort_boxplot}
\title{PLot bxoplot distribution of ranksum for a defined gene set according to ROI categories}
\usage{
cohort_boxplot(mixt.ranksum, tissue, module, cohort.name = "all",
  orderByTissue = NULL, orderByModule = NULL)
}
\arguments{
\item{mixt.ranksum}{Output of sig.ranksum()}

\item{tissue}{character string that provides the name of tissue we want to look at}

\item{module}{character string that provides the name of the module we want to generate a heatmap for}

\item{cohort.name}{character string that provides the name of the patient cohort we select patients from, defaults is set to 'all'}

\item{orderByTissue}{the tissue where the orderByModule module is found. Default is the same tissue.}

\item{orderByModule}{patient ordering can be based on another module/gene set expression.}
}
\description{
PLot bxoplot distribution of ranksum for a defined gene set according to ROI categories
}
