% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohort_scatterplot.R
\name{cohort_scatterplot}
\alias{cohort_scatterplot}
\title{Plot scatterplot of ranksums}
\usage{
cohort_scatterplot(mixt.ranksum, x.tissue, x.module, y.tissue, y.module,
  cohort.name = "all")
}
\arguments{
\item{mixt.ranksum}{output of sig.ranksum()}

\item{x.tissue}{character string providing the name of tissue on the x axis}

\item{x.module}{character string providing the name of module on the x axis}

\item{y.tissue}{character string providing the name of tissue on the y axis}

\item{y.module}{character string providing the name of module on the y axis}

\item{cohort.name}{character string that provides the name of the patient cohort.
Defaults is set to 'all'}
}
\description{
Generate scatterplot of ranksums obtained from two gene sets/modules
}
