% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functional.enrichment.R
\name{functional.enrichment}
\alias{functional.enrichment}
\title{Functional enrichment analyses for genes partitioned by sig.ranksum()}
\usage{
functional.enrichment(mixt.ranksum, tissue, cohort.name = "all",
  functional.groups, p.adjust.method = "BH", mc.cores = 2)
}
\arguments{
\item{mixt.ranksum}{Output of sig.ranksum()}

\item{tissue}{character string that provides the name of tissue of interest}

\item{cohort.name}{character string that provides the name of the patient
cohort to analyze, default is set to 'all'}

\item{functional.groups}{a list of length 2: each object is a named vector of
gene set appartenance for each tissue.}

\item{p.adjust.method}{correction method. See `?stats::p.adjust`}

\item{mc.cores}{number of cores to use}
}
\value{
The output is a list containing the following objects
\item{results}{data frame of enrichment results for each gene set/module.
Enrichment is estimated for all genes in the gene set/module and
for genes that are positively (up genes) or negatively (dn genes) correlated
with the patient ranksum. Number of genes in common with signature are
provided for all, up, and down gene groups.}
\item{updn.common}{names of genes in common between gene set/module and
functional signature.}
\item{up.common}{names of genes in common between up genes in gene set/module
and functional signature.}
\item{up.common}{names of genes in common between dn genes in gene set/module
and functional signature.}
}
\description{
Enrichemnt is estimated by the hypergeometric minimum-likelihood p-values,
computed with the function ‘dhyper’ (equivalent to one-sided Fisher
exact test). P-values are then adjusted for multiple testing.
}
