% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap.R
\name{heatmap.ddr}
\alias{heatmap.ddr}
\title{Flexible clustering function}
\usage{
heatmap.ddr(exprs = NULL, dist.elem = "cor", clust.elem = "average",
  p = 2)
}
\arguments{
\item{exprs}{a matrix-like object whose columns are to be clustered.
May be omitted if a dendrogram can be obtained using only the other arguments.}

\item{dist.elem}{specifies how to compute the distance matrix. May be
omitted if a dendrogram can be obtained from 'clust.elem' alone.
Possible specifications are: An object that can be coerced to a distance matrix
(using as.dist()), a function that takes 'exprs' and
produces a distance matrix, or a character string specifying
how a distance matrix is to be computed from 'exprs'.
The acceptable character strings are "cor" and any method specification that can be
passed on to the dist() function. If 'dist.elem' is "cor", then 1-correlation
(between the columns of 'exprs') is used as distance.}

\item{clust.elem}{specifies how to compute the dendrogram. Possible
specifications are:
An object that can be coerced to "dendrogram" (using as.dendrogram()), a function
that takes a distance matrix and produces an object that can be coerced to a dendrogram,
a character string specifying how to compute, from a distance
matrix, an object that can be coerced to a dendrogram, or a character
string specifying how to compute a dendrogram from a distance matrix.
passed to the dist() function and used only when 'dist.elem' is "minkowski"
(see help on the dist() function for details).}
}
\description{
\code{heatmap.ddr} is a very flexible function for clustering. It allows multiple ways of
specifying how the dendrogram is to be computed and it knows about
the cor() and hclust() functions.
}
\keyword{internal}
