% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap.R
\name{heatmap.is.valid.color.scheme}
\alias{heatmap.is.valid.color.scheme}
\title{Function for checking color schemes}
\usage{
heatmap.is.valid.color.scheme(color.scheme)
}
\description{
\code{heatmap.is.valid.color.scheme} internal function for checking color schemes.
}
\keyword{internal}
