% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap.R
\name{heatmap.key}
\alias{heatmap.key}
\title{Plot heatmap key}
\usage{
heatmap.key(x, color.scheme = heatmap.color.scheme(),
  key.min = utils::head(color.scheme$breaks, 1) - 1,
  key.max = utils::tail(color.scheme$breaks, 1) + 1,
  key.line.color = "cyan")
}
\arguments{
\item{x}{numeric matrix that was used to plot a heatmap}

\item{color.scheme}{the color scheme used to plot the heatmap}

\item{key.min}{should be less than the least break point of the color scheme}

\item{key.max}{should be greater than the greates break point of the color scheme}

\item{key.line.color}{the color used to draw the histogram}
}
\value{
invisible(NULL)
}
\description{
\code{heatmap.key} plots a heatmap key in the current viewport.
}
\keyword{internal}
