% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap.R
\name{heatmap.labels.cex}
\alias{heatmap.labels.cex}
\title{Return a cex value for text plotted in a viewport}
\usage{
heatmap.labels.cex(labels, type = c("row.labels", "col.labels"))
}
\arguments{
\item{labels}{character vector}

\item{type}{either "row.labels" or "col.labels" indicating whether the
labels are intended to be drawn horizontally or vertically.}
}
\value{
a cex value to be used when drawing the strings in 'labels' in
the current viewport
}
\description{
\code{heatmap.labels.cex}computes a cex value that can be used when plotting text in the
current viewport. The assumption is that 'labels' is a character
vector containing text strings that are meant to be drawn in the
current viewport (either horizontally or vertically). This function
will compute and return a cex value that can be passed to
grid.text() such that the strings in 'labels' will fit in the
current viewport when drawn on separate lines.
}
\keyword{internal}
