% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap.R
\name{heatmap.labels}
\alias{heatmap.labels}
\title{Plot labels in a viewport}
\usage{
heatmap.labels(labels, type = c("row.labels", "col.labels"),
  just = c("left", "right", "centre"), rot = 0)
}
\arguments{
\item{labels}{character vector with strings to be drawn in the current
viewport}

\item{type}{either "row.labels" or "col.lables", determining if the text
is to be drawn horizontally or vertically.}

\item{just}{the justification of the text.}

\item{rot}{rotation of the text in angles.}
}
\value{
invisible(NULL)
}
\description{
\code{heatmap.labels} plots each string in 'labels' on a separate line in the current
viewport. The size of the text and spacing between strings are
adjusted so that the text will fill the viewport as much as possible.
}
\keyword{internal}
