% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap.R
\name{heatmap.map}
\alias{heatmap.map}
\title{Plot heatmap}
\usage{
heatmap.map(x, color.scheme = heatmap.color.scheme(), na.color = "grey",
  col.grid = NA, use.raster = FALSE)
}
\arguments{
\item{x}{a numeric matrix with the heatmap values}

\item{color.scheme}{see the function heatmap.color.scheme() for details}

\item{na.color}{color used for NA values in 'x'}

\item{col.grid}{color used for grid lines in heatmap}
}
\value{
The color scheme invisibly.
}
\description{
\code{heatmap.map} plots a heatmap of 'x' in the current viewport. Returns the color
scheme used.
}
\keyword{internal}
