% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals.R
\name{hyper.fisher}
\alias{hyper.fisher}
\title{Hypergeometric test}
\usage{
hyper.fisher(pop1, pop2, bckg)
}
\description{
computation of hypergeometric test so it matches fisher exact test
}
\keyword{internal}
