% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.gmt.msigdb.R
\name{read.gmt.msigdb}
\alias{read.gmt.msigdb}
\title{Read MSigdb gmt files}
\usage{
read.gmt.msigdb(inputFile)
}
\arguments{
\item{inputFile}{ile character string pointing to the file to read}
}
\value{
list of msigdb gene sets
}
\description{
The gmt files downloaded from the MSigDB website are read and formatted
}
