% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sig.ranksum.R
\name{sig.ranksum}
\alias{sig.ranksum}
\title{Patient linear ordering based on gene set expression}
\usage{
sig.ranksum(x.dat, up = NULL, dn = NULL, ns = NULL, n = 1000,
  middle.range = 0.95, seed = 123456, mc.cores = 2)
}
\arguments{
\item{x.dat}{a list with at least 3 objects. exprs: gene expression data,
must be a matrix. Genes are in rows, patients in columns.
NAs in exprdata are not supported. clinical: clinical information
data frame. cohorts: list of patient subgroups}

\item{up}{indices of up-regulated genes}

\item{dn}{indices of down-regulated genes}

\item{ns}{indices of genes whose directions have not been specified.}

\item{n}{number of permutations to perform to compute the region of indendence}

\item{middle.range}{percentile point of the distribution of random patient
ranks to define the region of independence}

\item{seed}{random seed number}

\item{mc.cores}{number of cores to use}
}
\value{
The output is a list containing the following objects
\item{pat.order}{vector providing the patient linear ordering.
data[, pat.order] orders the
columns of
 expression matrix according to the sum of gene ranks (gene ranksums).}
\item{ranksum}{vector of gene ranksums acr}
m{gene.order}{gene ordering. exprdata[gene.order, ] sorts rows of exprdata
so that genes within each group around the two myeloids are ordered according
to the strength of their association with the patient linear ordering.
Most correlated genes are at each end of the
ordering}
\item{dat}{matrix of expression of \code{up and dn} or \code{ns} genes
reordered by patient and gene ordering.}
\item{up.dn}{vector with nrow(dat) values. -1, indicates down-gene
       and 1 indicates up-gene.}
\item{up}{vector of indexes pointing at M1 genes in exprdata defined as upregulated
i.e positively correlated with patient ordering.}
\item{dn}{vector of indexes pointing at M2 genes in exprdata defined as downregulated
i.e negatively correlated with patient ordering.}
}
\description{
\code{sig.ranksum} linearly orders samples based on expression of a defined
set of genes. Patients with the same ranksum are assigned their 'average' rank,
i.e., they all receive the same rank value, which might not be an integer.

Genes in the set are partitioned into two groups around myeloids (ğ‘€1 and ğ‘€2)
using correlation as the distance metric. Each gene in ğ‘€1 and ğ‘€2 is ordered
from high to low and low to high expression, respectively.
Expression values of each gene are then replaced by their ranks across patients.
The sum of gene ranks (ranksum) are then used to linearly ordered patients.
Note that 'up' and 'dn' or 'ns' are used to index into exprdata.
}
