% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat.ranksum.R
\name{stat.ranksum}
\alias{stat.ranksum}
\title{Compute association between gene sets/modules across tissues}
\usage{
stat.ranksum(mixt.ranksum, tissue1, tissue2, corType = "p", nRuns = 10000,
  randomSeed = 12345, mc.cores = 2, verbose = 2)
}
\arguments{
\item{mixt.ranksum}{output of sig.ranksum()}

\item{tissue1}{name of the first tissue the test is performed for.
should be a valid name of mixt.ranksum}

\item{tissue2}{name of second tissue the test is performed for.
should be a valid name of mixt.ranksum.}

\item{corType}{a character string indicating which correlation coefficient
is to be computed.
Default 'p' for pearson}

\item{nRuns}{number of permutations}

\item{randomSeed}{seed number for random number generation.
Default set as '1234'}

\item{mc.cores}{number of cores}

\item{verbose}{numerical. default > 0 show informational text on progress}
}
\description{
Interactions between modules across tissues are identified using a random permutation approach
based on the correlation between ranksums of gene expression in gene sets/modules across tissues.
}
